refactor email into its own package
