+++ univocity-trader-core/src/main/java/com/univocity/trader/account/Client.java
@@ -11,9 +11,9 @@ import com.univocity.trader.ClientAccount;
 import com.univocity.trader.SymbolPriceDetails;
 import com.univocity.trader.candles.Candle;
 import com.univocity.trader.candles.CandleProcessor;
+import com.univocity.trader.email.SmtpMailSender;
 import com.univocity.trader.exchange.Exchange;
 import com.univocity.trader.notification.OrderExecutionToEmail;
-import com.univocity.trader.notification.SmtpMailSender;
 import com.univocity.trader.simulation.DefaultConfiguration;
 import com.univocity.trader.simulation.Parameters;
 import com.univocity.trader.strategy.Engine;
