+++ univocity-trader-core/src/main/java/com/univocity/trader/email/Email.java
@@ -5,7 +5,7 @@
  *
  */
 
-package com.univocity.trader.notification;
+package com.univocity.trader.email;
 
 public class Email {
