+++ univocity-trader-core/src/main/java/com/univocity/trader/email/MailSenderConfig.java
@@ -1,4 +1,4 @@
-package com.univocity.trader.notification;
+package com.univocity.trader.email;
 
 
 import org.apache.commons.lang3.*;
